TodoListDetailView -> TodoListItemView
